getMoney 반환값 long 수정

+++ src/main/java/com/bitcamp/project/dao/TradeDAO.java
@@ -23,7 +23,7 @@ public interface TradeDAO {
 
 	public void clearChart();
 
-	public int getMoney(String id);
+	public long getMoney(String id);
 	
 	public int getStockQuantity(StockVO vo);
 }
